ErrorManager: direct copy effectiveMessageCode
